CadastroCliente/Services/ClienteService.java
package Services;

import Scr.main.java.Cliente;
import Scr.main.java.DAO.IClienteDAO;
import Scr.main.java.DAO.Generics.Exeption.TipoChaveNaoEncontradaException;

public class ClienteService implements IClienteService {

    private IClienteDAO clienteDAO;

    public ClienteService(IClienteDAO clienteDAO){
        this.clienteDAO = clienteDAO;
    }

    @Override
    public boolean cadastrar(Cliente cliente) throws TipoChaveNaoEncontradaException{
    return clienteDAO.cadastrar(cliente);
    }

    @Override
    public Cliente buscarPorCPF(Long cpf) {
    return clienteDAO.consultar(cpf);
    }

    @Override
    public void excluir(Long cpf) {
        clienteDAO.excluir(cpf);
    }

    @Override
    public void alterar(Cliente cliente) throws TipoChaveNaoEncontradaException {
        clienteDAO.alterar(cliente);
    }

}
